Fix di alcuni bug e implementazione commands e queries

+++ src/BlaisePascal.SmartHouse.Domain/Interfaces/IGetConsumption.cs
@@ -8,6 +8,6 @@ namespace BlaisePascal.SmartHouse.Domain.Interfaces
 {
     public interface IGetConsumption
     {
-        int GetConsumption();
+        float GetConsumption();
     }
 }
